Task 0 updated

+++ 0x1E-search_algorithms/0-linear.c
@@ -2,26 +2,26 @@
 #include <stdio.h>
 
 /**
- * linear_search - searches for a value in an array of integers
- *                 using the Linear search algorithm
- * @array: pointer to the first element of the array to search in
- * @size: number of elements in array
- * @value: value to search for
- * Return: Index where val located, or -1 if val not present or array is NULL
+ * linear_search - Searches for a value in an array using Linear search
+ * @array: Pointer to the first element of the array
+ * @size: Number of elements in the array
+ * @value: Value to search for
+ *
+ * Return: First index where value is found, or -1 if not found
  */
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i;
+size_t i;
 
-	if (array == NULL)
-		return (-1);
+if (array == NULL)
+return (-1);
 
-	for (i = 0; i < size; i++)
-	{
-		printf("Value checked array[%lu] = [%d]\n", i, array[i]);
-		if (array[i] == value)
-			return (i);
-	}
+for (i = 0; i < size; i++)
+{
+printf("Value checked array[%lu] = [%d]\n", i, array[i]);
+if (array[i] == value)
+return ((int)i);
+}
 
-	return (-1);
+return (-1);
 }
